pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task management is the process of managing a task through its lifecycle. It involves planning, testing, tracking."
authors = ["Sergei Mikurov <454004@mail.ru>"]
readme = "README.md"
repository = "https://github.com/serVmik/python-project-52.git"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Other Audience",
    "License :: Freeware",
    "Natural Language :: English, Russian",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
]

packages = [
    { include = "task_manager" }
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
python-dotenv = "^1.0.0"
django-bootstrap5 = "^23.3"
python-json-logger = "^2.0.7"
dj-database-url = "^2.1.0"
whitenoise = "^6.5.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-filter = "^23.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.2"
coverage = "^7.3.1"
pytest-cov = "^4.1.0"
rollbar = "^0.16.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
